fix: repeated logs saved in the local storage

+++ src/reducers/settings.js
@@ -65,21 +65,24 @@ export default (state, action) => {
     case SET_CURRENT_SESSION:
       return { ...state, currentSession: action.payload }
     case SET_END_TIME_AND_SAVE:
-      const clonedObj = Object.assign(state.currentSession)
-      clonedObj.endTime = action.payload // add end time
-      let userLogs = []
-      //TODO
-      // MAKE SURE TO CHECK FOR UNIQUE LOGS BY ID
-      // AT THIS MOMENT IS BEING ADDED 3 IDENTICAL INSTACES OF THE SAME LOG. WHY?????
-
-      // Parse the serialized data back into an aray of objects
-      userLogs = JSON.parse(localStorage.getItem('session')) || []
-      // Push the new data (whether it be an object or anything else) onto the array
-      userLogs.push(clonedObj)
-      // Re-serialize the array back into a string and store it in localStorage
-      localStorage.setItem('session', JSON.stringify(userLogs))
-      // Reset current session on the state
-      return { ...state, currentSession: clonedObj }
+      const ls = localStorage.session && JSON.parse(localStorage.session)
+      const isLogExists =
+        ls && ls.find((session) => session.id === state.currentSession.id)
+      if (!isLogExists) {
+        const clonedObj = Object.assign(state.currentSession)
+        // add end time
+        clonedObj.endTime = action.payload
+        let userLogs = []
+        // Parse the serialized data back into an aray of objects
+        userLogs = JSON.parse(localStorage.getItem('session')) || []
+        // Push the new data (whether it be an object or anything else) onto the array
+        userLogs.push(clonedObj)
+        // Re-serialize the array back into a string and store it in localStorage
+        localStorage.setItem('session', JSON.stringify(userLogs))
+        // Reset current session on the state
+        return { ...state, currentSession: clonedObj }
+      }
+      return { ...state }
     default:
       return state
   }
